+++ src/scoresaber/rankeds.js
@@ -1,0 +1,3 @@
+import {getCacheAndConvertIfNeeded} from "../store";
+
+export const getRankedMaps = async _ => (await getCacheAndConvertIfNeeded()).rankedSongs;