+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 0.7.3
- - refresh component added to player profile and global country ranking
+ - refresh component added to player profile
  
 0.7.2
  - add pp change on global country leaderboard